Move app/run.go to cmd/server/run.go from initialized app
