+++ pkg/grapicmd/internal/module/generator/.snapshots/TestProjectGenerator_GenerateProject-no_config-cmd-server-main.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"os"
+
+	"google.golang.org/grpc/grpclog"
+)
+
+func main() {
+	err := run()
+	if err != nil {
+		grpclog.Errorf("server was shutdown with errors: %v", err)
+		os.Exit(1)
+	}
+}
+
