--- pkg/grapicmd/internal/module/generator/.snapshots/TestProjectGenerator_GenerateProject-with_UseHEAD-app-run.go
@@ -1,17 +1,0 @@
-package app
-
-import (
-	"github.com/izumin5210/grapi/pkg/grapiserver"
-)
-
-// Run starts the grapiserver.
-func Run() error {
-	s := grapiserver.New(
-		grapiserver.WithDefaultLogger(),
-		grapiserver.WithServers(
-		// TODO
-		),
-	)
-	return s.Serve()
-}
-
