+++ pkg/grapicmd/internal/module/generator/template/init/cmd/server/main.go.tmpl
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"os"
+
+	"google.golang.org/grpc/grpclog"
+)
+
+func main() {
+	err := run()
+	if err != nil {
+		grpclog.Errorf("server was shutdown with errors: %v", err)
+		os.Exit(1)
+	}
+}
