+++ pkg/grapicmd/internal/module/generator/template/init/cmd/server/run.go.tmpl
@@ -1,22 +1,15 @@
 package main
 
 import (
-	"os"
-
-	"google.golang.org/grpc/grpclog"
-
-	"{{ .importPath }}/app"
+	"github.com/izumin5210/grapi/pkg/grapiserver"
 )
 
-func main() {
-	os.Exit(run())
-}
-
-func run() int {
-	err := app.Run()
-	if err != nil {
-		grpclog.Errorf("server was shutdown with errors: %v", err)
-		return 1
-	}
-	return 0
+func run() error {
+	s := grapiserver.New(
+		grapiserver.WithDefaultLogger(),
+		grapiserver.WithServers(
+		// TODO
+		),
+	)
+	return s.Serve()
 }
